add custclaim command to sweep a custCloseTx
New cli CustClaim -  given an escrowTxid for a closed channel, it will broadcast the claimTx to sweep the output into it's own address.

storeCustClaimTx is the main function in zkChainWatcher that signs and stores the custClaimTx. For now, the output Pubkey is the same as the custPubkey in custCloseTx.

custClaimTx is stored in zkClaimDB, under a bucket with escrowTxid as the key.

CustClaim in zkChannelManager loads and broadcasts the claimTx

This commit also includes updates to work with the updated libzkchannels